+++ lib/dotgen/class2.c
@@ -67,9 +67,7 @@ static node_t *leader_of(node_t * v) {
     return rv;
 }
 
-/* make_chain:
- * Create chain of dummy nodes for edge orig.
- */
+/// create chain of dummy nodes for edge orig
 static void 
 make_chain(graph_t * g, node_t * from, node_t * to, edge_t * orig)
 {