+++ lib/dotgen/decomp.c
@@ -72,8 +72,7 @@ static node_t *pop(node_stack_t *sp) {
   return node_stack_pop_back(sp);
 }
 
-/* search_component:
- * iterative dfs for components.
+/* iterative dfs for components.
  * We process the edges in reverse order of the recursive version to maintain
  * the processing order of the nodes.
  * Since are using a stack, we need to indicate nodes on the stack. Nodes unprocessed
@@ -103,7 +102,7 @@ static void search_component(node_stack_t *stk, graph_t *g, node_t *n) {
 		    e = *ep;
 		    if ((other = aghead(e)) == n)
 			other = agtail(e);
-		    if ((ND_mark(other) != Cmark) && (other == UF_find(other)))
+		    if (ND_mark(other) != Cmark && other == UF_find(other))
 			push(stk, other);
 		}
 	    }
@@ -122,7 +121,7 @@ void decompose(graph_t * g, int pass)
     GD_comp(g).size = 0;
     for (n = agfstnode(g); n; n = agnxtnode(g, n)) {
 	v = n;
-	if ((pass > 0) && (subg = ND_clust(v)))
+	if (pass > 0 && (subg = ND_clust(v)))
 	    v = GD_rankleader(subg)[ND_rank(v)];
 	else if (v != UF_find(v))
 	    continue;